Enforce reduction of 3D geometry to 2D
Most underlying functions (e.g. distance measurement in shapely) do
not support 3D geometries. Having z-coordinates in the
data can cause issues in several places in the code, especially when
e.g. buildings contain z coordinates, but streets do not.

The savest option currently is to force 2d on all input geometries,
which is now the default in check_crs()