Allow multiple potential building connections
DHNx requires connecting each building to the line network, e.g.
the streets. To do this, create_object_connections() normally
finds the perpendicular line which represents the closest
connection between building and street.

This commit updates create_object_connections() to allow more than
1 (default value) number of connections 'n_conn' to be created.
During the optimization phase, the solver now has multiple options
to choose from when finding the best district heating network.

This may be desired when the nearest street segment is not actually
the best way to connect a building. In cases where a building has
a street or path in the front and back, both may be valid options.

For each line segment that a possible connection has been attached
to, its direct connections are ignored when searching for the next
possible line segment. This mostly avoids obviously inferior
alternative connections and helps find more relevant alternatives.

To allow this functionality, create_object_connections() had
to be completely rewritten. But the default settings should
retain the old behaviour.

This feature increases the optimization time.